вынесла вспомогательные функции в отдельный модуль util.js, функции генерации данных в модуль data.js
