+++ index.html
@@ -11,7 +11,6 @@
 
 <body>
   <main>
-
     <!-- Контейнер для изображений от других пользователей -->
     <section class="pictures  container">
       <h2 class="pictures__title  visually-hidden">Фотографии других пользователей</h2>
@@ -164,6 +163,6 @@
       </div>
     </section>
   </template>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
